+++ src/CMakeLists.txt
@@ -214,6 +214,7 @@ add_library(parthenon
   utils/string_utils.cpp
   utils/string_utils.hpp
   utils/utils.hpp
+  utils/variadic_template_utils.hpp
 
   argument_parser.hpp
   basic_types.hpp
